requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
PyYAML==6.0.2
langgraph==0.2.39
typer==0.12.5
click==8.1.7
pytest==8.3.3
google-genai>=1.0.0
